requirements.txt
fastapi
uvicorn
python-multipart
openai
aiofiles
httpx
pyyaml
